Fix DOMSelection.addRange issue and expose DOMRange and DOMSelection
DOMSelection.addRange doesn't work correctly if parentElement is
contentEditable.
Because DOMRange._selectionExists return true if parentElement is
contentEditable for getting cursor position but in this case the
function should return false. So I add `cursor` argument and if cursor
argument set `false` then return false while no selection is exists even
if parentElement is contentEditable.

Exposing DOMRange and DOMSelection is required to use these class in
iframe. So I add DOMRange and DOMSelection to window.

Previous code, even window.getSelection is defined, window.getSelection
is overwrite with DOMSelection so i fixed to not overwrite it as well.